+++ src/test/java/io/github/astrapi69/crypt/data/example/keyagreement/KeyAgreementTest.java
@@ -26,10 +26,14 @@ package io.github.astrapi69.crypt.data.example.keyagreement;
 
 import static javax.xml.bind.DatatypeConverter.printHexBinary;
 
+import java.security.InvalidKeyException;
 import java.security.KeyFactory;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
+import java.security.NoSuchAlgorithmException;
 import java.security.PublicKey;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.KeySpec;
 import java.security.spec.X509EncodedKeySpec;
 
 import javax.crypto.Cipher;
@@ -37,10 +41,13 @@ import javax.crypto.KeyAgreement;
 import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.DESKeySpec;
+import javax.crypto.spec.SecretKeySpec;
 
 import org.junit.jupiter.api.Test;
 
 import io.github.astrapi69.crypt.data.factory.KeyPairGeneratorFactory;
+import io.github.astrapi69.crypt.data.factory.KeySpecFactory;
+import io.github.astrapi69.crypt.data.factory.SecretKeyFactoryExtensions;
 
 /**
  * The unit test for an example for KeyAgreement communication
@@ -53,13 +60,24 @@ public class KeyAgreementTest
 	@Test
 	void newSharedSecret() throws Exception
 	{
+		String cipherAlgorithm;
+		String secretKeyAlgorithm;
+		String keyAgreementAlgorithm;
+		String keyFactoryAlgorithm;
+		String keyPairGeneratorAlgorithm;
+		KeyPairGenerator keyPairGenerator;
+		int keySize;
+
 		// 1. The initiator(Alice in our case) generates a key pair(public and private) and sends
 		// the public key, along with the algorithm specification, to the other(Bob in our case)
 		// party.
-		KeyPairGenerator keyPairGenerator;
-		String algorithm = "EC";
-		int keySize = 256;
-		keyPairGenerator = KeyPairGeneratorFactory.newKeyPairGenerator(algorithm);
+		keyPairGeneratorAlgorithm = "EC";
+		keyFactoryAlgorithm = keyPairGeneratorAlgorithm;
+		secretKeyAlgorithm = "DES";
+		keyAgreementAlgorithm = "ECDH";
+		cipherAlgorithm = "DES/ECB/PKCS5Padding";
+		keySize = 256;
+		keyPairGenerator = KeyPairGeneratorFactory.newKeyPairGenerator(keyPairGeneratorAlgorithm);
 		keyPairGenerator.initialize(keySize);
 
 		KeyPair aliceKeyPair = keyPairGenerator.generateKeyPair();
@@ -71,7 +89,7 @@ public class KeyAgreementTest
 		// 2. The other(Bob in our case) party generates its own key pair(public and private) using
 		// the algorithm specification and sends the public key to the initiator(Alice in our case).
 
-		keyPairGenerator = KeyPairGeneratorFactory.newKeyPairGenerator(algorithm);
+		keyPairGenerator = KeyPairGeneratorFactory.newKeyPairGenerator(keyPairGeneratorAlgorithm);
 		keyPairGenerator.initialize(keySize);
 		KeyPair bobKeyPair = keyPairGenerator.generateKeyPair();
 
@@ -79,13 +97,13 @@ public class KeyAgreementTest
 		// Display public key from Bob
 		System.out.format("Bob Public Key: %s%n", printHexBinary(bobPk));
 
-		KeyFactory keyFactory = KeyFactory.getInstance("EC");
+		KeyFactory keyFactory = KeyFactory.getInstance(keyFactoryAlgorithm);
 		X509EncodedKeySpec bobPkSpec = new X509EncodedKeySpec(bobPk);
 		PublicKey bobPublicKey = keyFactory.generatePublic(bobPkSpec);
 		// 3. The initiator(Alice in our case) generates the secret key using its private key and
 		// the other(Bob in our case) party's public key.
 		// Create key agreement
-		KeyAgreement aliceKeyAgreement = KeyAgreement.getInstance("ECDH");
+		KeyAgreement aliceKeyAgreement = KeyAgreement.getInstance(keyAgreementAlgorithm);
 		aliceKeyAgreement.init(aliceKeyPair.getPrivate());
 		aliceKeyAgreement.doPhase(bobPublicKey, true);
 
@@ -97,7 +115,7 @@ public class KeyAgreementTest
 		// initiator's public key. Diffie-Hellamn algorithm ensures that both parties generate the
 		// same secret key.
 		// Create key agreement for Bob
-		KeyAgreement bobKeyAgreement = KeyAgreement.getInstance("ECDH");
+		KeyAgreement bobKeyAgreement = KeyAgreement.getInstance(keyAgreementAlgorithm);
 		bobKeyAgreement.init(bobKeyPair.getPrivate());
 		bobKeyAgreement.doPhase(aliceKeyPair.getPublic(), true);
 
@@ -106,25 +124,35 @@ public class KeyAgreementTest
 		System.out.format("Bob Shared secret: %s%n", printHexBinary(bobSharedSecret));
 
 		// 5. generate key from Alice shared secret
-		SecretKeyFactory aliceSecretKeyFactory = SecretKeyFactory.getInstance("DES");
-		DESKeySpec aliceKeySpec = new DESKeySpec(aliceSharedSecret);
-		SecretKey aliceSecretKey = aliceSecretKeyFactory.generateSecret(aliceKeySpec);
+		SecretKey aliceSecretKey = SecretKeyFactoryExtensions.newSecretKey(aliceSharedSecret,
+			secretKeyAlgorithm);
 
-		Cipher aliceCipher = Cipher.getInstance("DES/ECB/PKCS5Padding");
+		Cipher aliceCipher = Cipher.getInstance(cipherAlgorithm);
 		aliceCipher.init(Cipher.ENCRYPT_MODE, aliceSecretKey);
 		byte[] encryptedByteArray = aliceCipher
 			.doFinal("Free your mind and get the best out of you".getBytes());
 
 		// Step 6: generate key from Bob shared secret
-		SecretKeyFactory bobSecretKeyFactory = SecretKeyFactory.getInstance("DES");
+		SecretKeyFactory bobSecretKeyFactory = SecretKeyFactory.getInstance(secretKeyAlgorithm);
 		DESKeySpec bobKeySpec = new DESKeySpec(bobSharedSecret);
 		SecretKey bobSecretKey = bobSecretKeyFactory.generateSecret(bobKeySpec);
 
 		// Step 8: Bob receives the encrypted text and decrypts it
-		Cipher bobCipher = Cipher.getInstance("DES/ECB/PKCS5Padding");
+		Cipher bobCipher = Cipher.getInstance(cipherAlgorithm);
 		bobCipher.init(Cipher.DECRYPT_MODE, bobSecretKey);
 		byte[] plaintext = bobCipher.doFinal(encryptedByteArray);
 		String text = new String(plaintext);
 		System.out.println("Bob reads the message of Alice:\n" + text);
+	}
+
+
+	public static SecretKey newSecretKey(byte[] sharedSecret, String secretKeyAlgorithm)
+		throws NoSuchAlgorithmException, InvalidKeySpecException
+	{
+		SecretKeyFactory secretKeyFactory = SecretKeyFactory.getInstance(secretKeyAlgorithm);
+		SecretKeySpec secretKeySpec = KeySpecFactory.newSecretKeySpec(sharedSecret,
+			secretKeyAlgorithm);
+		SecretKey secretKey = secretKeyFactory.generateSecret(secretKeySpec);
+		return secretKey;
 	}
 }